UnsupportedObjectTypeInTriplet and UnrelatedMappingException moved
